:recycle: api: rename complementary certif API file
