+++ api/tests/certification/complementary-certification/integration/application/api/complementary-certification-api_test.js
@@ -1,4 +1,4 @@
-import { getById } from '../../../../../../src/certification/complementary-certification/application/api/complementary-api.js';
+import { getById } from '../../../../../../src/certification/complementary-certification/application/api/complementary-certification-api.js';
 import { catchErr, databaseBuilder, expect } from '../../../../../test-helper.js';
 import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
 import { ComplementaryCertification } from '../../../../../../src/certification/complementary-certification/application/api/models/ComplementaryCertification.js';
